Feature batch, refresh fix take 2
Added color option to set the color of the warning text

When the warning text is an empty string, the widget has 0 height,
effectively hiding it.

Added a warn type as well as options for including nodes of type
'deprecated' or 'experimental' in search anyways. If a widget does not
have an assigned type, it defaults to deprecated.

Actually fix nodes being hidden on refresh. The last attempt was merely
nothing ever being hidden because of a syntax error preventing teh
script from loading. This implementation takes the rather more heavy
handed approach of applying at beforeRegisterNodeDef and setting up the
global filter option at init. This makes it marginally more likely to
interfere with other plugins that attempt to modify app.graph.filter, so
added safety has been included elsewhere in the code.

Added example usage to to Inject PIA as to demonstrate how it might be
used for an experimental node.