widget selection, collapse allignment, video style
Mouse selection of individual widgets now correctly applies for nodes
which do not contain a domWidget and therefore lack a y attribute

The collapse button has been migrated to use a monospace font for the +
and minus characters to ensure a change in state doesn't cause a change
in character alignment. The js code has been update to account for this
change

Basic styling has been applied for embedded videos. A play and pause
option are provided, but the mute and fullscreen buttons are hidden.
This functionality is tied to a class applied by documentation.py and
can easily be removed if a video embed is used for more advanced
information (a tutorial) instead of just quick demonstrations or output.