Shrink width of long title nodes by default
Because of a change in initialization order. ComfyUI now considers a
node's title before scaling by 1.5 to determine a nodes initial width.
This causes nodes with long titles to have correspondingly large
dead space. As a rough fix, ADE nodes are changed to have initial size
equal to the greater of the default node width or the minimum title
length.