Remove old deprecation code.
ComfyUI now has native support for marking nodes as deprecated or
experimental, so the much of the code here no longer serves a purpose.

Zombie code for hiding nodes has been removed and this code now only
exists to display a text label on the node of a specific color.

This removes excessive console spam about deprecation in the reading of
setting values.