Fix documentation scrollbar style leakage
The floating documentation applies styling to scrollbars to make them
take up less space and better match the styling of the canvas.

This styling is improperly configured and will leak to other areas in
comryui with scrollabrs such as sidebar.

This style code is only applied if a user has installed ADE but not VHS.